Convert PythonVirtualEnv Document from Markdown to Text.
Majority of the time, documentation is read by users and developers
from the command-line either using a pager (less, more, ...) or an
editor (EMACS, vi, ...).  For this reason it is better to format the
documentation as plain text than to have it in markdown format.
Markdown makes it difficult to read using these tools.  It also
requires a tool to render it.

This change updates the PythonVirtualEnv.md document and converts it
to plain text.